@@ -1,28 +1,30 @@
 {
-  "_from": "@firebase/app-types@0.3.2",
-  "_id": "@firebase/app-types@0.3.2",
+  "_from": "@firebase/app-types@0.6.0",
+  "_id": "@firebase/app-types@0.6.0",
   "_inBundle": false,
-  "_integrity": "sha512-ZD8lTgW07NGgo75bTyBJA8Lt9+NweNzot7lrsBtIvfciwUzaFJLsv2EShqjBeuhF7RpG6YFucJ6m67w5buCtzw==",
+  "_integrity": "sha512-ld6rzjXk/SUauHiQZJkeuSJpxIZ5wdnWuF5fWBFQNPaxsaJ9kyYg9GqEvwZ1z2e6JP5cU9gwRBlfW1WkGtGDYA==",
   "_location": "/@firebase/app-types",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@firebase/app-types@0.3.2",
+    "raw": "@firebase/app-types@0.6.0",
     "name": "@firebase/app-types",
     "escapedName": "@firebase%2fapp-types",
     "scope": "@firebase",
-    "rawSpec": "0.3.2",
+    "rawSpec": "0.6.0",
     "saveSpec": null,
-    "fetchSpec": "0.3.2"
+    "fetchSpec": "0.6.0"
   },
   "_requiredBy": [
-    "/@firebase/app"
+    "/@firebase/app",
+    "/@firebase/database-types",
+    "/firebase"
   ],
-  "_resolved": "https://registry.npmjs.org/@firebase/app-types/-/app-types-0.3.2.tgz",
-  "_shasum": "a92dc544290e2893bd8c02a81e684dae3d8e7c85",
-  "_spec": "@firebase/app-types@0.3.2",
-  "_where": "/Users/muzzicm/myStuff/chrismuzzin.info/functions/node_modules/@firebase/app",
+  "_resolved": "https://registry.npmjs.org/@firebase/app-types/-/app-types-0.6.0.tgz",
+  "_shasum": "8dcc3e793c6983e9d54f7eb623a7618c05f2d94c",
+  "_spec": "@firebase/app-types@0.6.0",
+  "_where": "C:\\Users\\dirty\\workspaces\\Website\\functions\\node_modules\\firebase",
   "author": {
     "name": "Firebase",
     "email": "firebase-support@google.com",
@@ -35,20 +37,22 @@
   "deprecated": false,
   "description": "@firebase/app Types",
   "devDependencies": {
-    "typescript": "2.8.1"
+    "typescript": "3.8.3"
   },
   "files": [
     "index.d.ts",
     "private.d.ts"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "name": "@firebase/app-types",
   "repository": {
+    "directory": "packages/app-types",
     "type": "git",
-    "url": "https://github.com/firebase/firebase-js-sdk/tree/master/packages/app-types"
+    "url": "git+https://github.com/firebase/firebase-js-sdk.git"
   },
   "scripts": {
     "test": "tsc"
   },
-  "version": "0.3.2"
+  "version": "0.6.0"
 }
