@@ -34,48 +34,31 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "@firebase/app-types": "0.3.2",
-    "@firebase/util": "0.2.2",
+    "@firebase/app-types": "0.6.0",
+    "@firebase/component": "0.1.8",
+    "@firebase/logger": "0.2.0",
+    "@firebase/util": "0.2.43",
     "dom-storage": "2.1.0",
-    "tslib": "1.9.0",
+    "tslib": "1.11.1",
     "xmlhttprequest": "1.8.0"
   },
   "deprecated": false,
   "description": "The primary entrypoint to the Firebase JS SDK",
   "devDependencies": {
-    "@types/chai": "4.1.2",
-    "@types/mocha": "5.0.0",
-    "@types/sinon": "4.3.1",
-    "chai": "4.1.2",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-firefox-launcher": "1.1.0",
-    "karma-mocha": "1.3.0",
-    "karma-sauce-launcher": "1.2.0",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "2.0.9",
-    "mocha": "5.0.5",
-    "npm-run-all": "4.1.2",
-    "nyc": "11.6.0",
-    "rollup": "0.57.1",
-    "rollup-plugin-commonjs": "9.1.0",
-    "rollup-plugin-node-resolve": "3.3.0",
-    "rollup-plugin-replace": "2.0.0",
-    "rollup-plugin-typescript2": "0.12.0",
-    "sinon": "4.5.0",
-    "source-map-loader": "0.2.3",
-    "ts-loader": "3.5.0",
-    "ts-node": "5.0.1",
-    "typescript": "2.8.1",
-    "webpack": "3.11.0"
+    "rollup": "2.0.6",
+    "rollup-plugin-json": "4.0.0",
+    "rollup-plugin-replace": "2.2.0",
+    "rollup-plugin-typescript2": "0.26.0",
+    "typescript": "3.8.3"
   },
+  "esm2017": "dist/index.esm2017.js",
   "files": [
     "dist"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
+  "lite": "dist/index.lite.js",
+  "lite-esm2017": "dist/index.lite.esm2017.js",
   "main": "dist/index.node.cjs.js",
   "module": "dist/index.esm.js",
   "name": "@firebase/app",
