@@ -1,28 +1,28 @@
 {
-  "_from": "@firebase/auth-types@0.3.4",
-  "_id": "@firebase/auth-types@0.3.4",
+  "_from": "@firebase/auth-types@0.10.0",
+  "_id": "@firebase/auth-types@0.10.0",
   "_inBundle": false,
-  "_integrity": "sha512-0r3gSQk9jw5orFHCTUIgao0zan6dHt2J0BO3t/uEzbod+uwqvUn/gh+yg+kK6HX92Fg8E7y030KX4Bw/aXt0Ew==",
+  "_integrity": "sha512-VuW7c+RAk3AYPU0Hxmun3RzXn7fbJDdjQbxvvpRMnQ9zrhk8mH42cY466M0n4e/UGQ+0smlx5BqZII8aYQ5XPg==",
   "_location": "/@firebase/auth-types",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@firebase/auth-types@0.3.4",
+    "raw": "@firebase/auth-types@0.10.0",
     "name": "@firebase/auth-types",
     "escapedName": "@firebase%2fauth-types",
     "scope": "@firebase",
-    "rawSpec": "0.3.4",
+    "rawSpec": "0.10.0",
     "saveSpec": null,
-    "fetchSpec": "0.3.4"
+    "fetchSpec": "0.10.0"
   },
   "_requiredBy": [
     "/@firebase/auth"
   ],
-  "_resolved": "https://registry.npmjs.org/@firebase/auth-types/-/auth-types-0.3.4.tgz",
-  "_shasum": "253b1b2d9b520a0b945d4617c8418f0f19a4159f",
-  "_spec": "@firebase/auth-types@0.3.4",
-  "_where": "C:\\Users\\dirty\\workspaces\\chrismuzzin.info\\functions\\node_modules\\@firebase\\auth",
+  "_resolved": "https://registry.npmjs.org/@firebase/auth-types/-/auth-types-0.10.0.tgz",
+  "_shasum": "9403633e723336055fad4bbf5e4c9fe3c55f8d3f",
+  "_spec": "@firebase/auth-types@0.10.0",
+  "_where": "C:\\Users\\dirty\\workspaces\\Website\\functions\\node_modules\\@firebase\\auth",
   "author": {
     "name": "Firebase",
     "email": "firebase-support@google.com",
@@ -35,22 +35,25 @@
   "deprecated": false,
   "description": "@firebase/auth Types",
   "devDependencies": {
-    "typescript": "2.8.1"
+    "typescript": "3.8.3"
   },
   "files": [
     "index.d.ts"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "name": "@firebase/auth-types",
   "peerDependencies": {
-    "@firebase/app-types": "0.x"
+    "@firebase/app-types": "0.x",
+    "@firebase/util": "0.x"
   },
   "repository": {
+    "directory": "packages/auth-types",
     "type": "git",
-    "url": "https://github.com/firebase/firebase-js-sdk/tree/master/packages/auth-types"
+    "url": "git+https://github.com/firebase/firebase-js-sdk.git"
   },
   "scripts": {
     "test": "tsc"
   },
-  "version": "0.3.4"
+  "version": "0.10.0"
 }
