@@ -1,28 +1,28 @@
 {
-  "_from": "@firebase/database-types@0.3.2",
-  "_id": "@firebase/database-types@0.3.2",
+  "_from": "@firebase/database-types@0.4.14",
+  "_id": "@firebase/database-types@0.4.14",
   "_inBundle": false,
-  "_integrity": "sha512-9ZYdvYQ6r3aaHJarhUM5Hf6lQWu3ZJme+RR0o8qfBb9L04TL3uNjt+AJFku1ysVPntTn+9GqJjiIB2/OC3JtwA==",
+  "_integrity": "sha512-+D41HWac0HcvwMi+0dezEdSOZHpVjPKPNmpQiW2GDuS5kk27/v1jxc9v7F4ALLtpxbVcn16UZl5PqEkcS9H2Xg==",
   "_location": "/@firebase/database-types",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@firebase/database-types@0.3.2",
+    "raw": "@firebase/database-types@0.4.14",
     "name": "@firebase/database-types",
     "escapedName": "@firebase%2fdatabase-types",
     "scope": "@firebase",
-    "rawSpec": "0.3.2",
+    "rawSpec": "0.4.14",
     "saveSpec": null,
-    "fetchSpec": "0.3.2"
+    "fetchSpec": "0.4.14"
   },
   "_requiredBy": [
     "/@firebase/database"
   ],
-  "_resolved": "https://registry.npmjs.org/@firebase/database-types/-/database-types-0.3.2.tgz",
-  "_shasum": "70611a64dd460e0e253c7427f860d56a1afd86fe",
-  "_spec": "@firebase/database-types@0.3.2",
-  "_where": "/Users/muzzicm/myStuff/chrismuzzin.info/functions/node_modules/@firebase/database",
+  "_resolved": "https://registry.npmjs.org/@firebase/database-types/-/database-types-0.4.14.tgz",
+  "_shasum": "181e10c1d1ae64fd0a080f6e0369cec115c51d70",
+  "_spec": "@firebase/database-types@0.4.14",
+  "_where": "C:\\Users\\dirty\\workspaces\\Website\\functions\\node_modules\\@firebase\\database",
   "author": {
     "name": "Firebase",
     "email": "firebase-support@google.com",
@@ -32,25 +32,27 @@
     "url": "https://github.com/firebase/firebase-js-sdk/issues"
   },
   "bundleDependencies": false,
+  "dependencies": {
+    "@firebase/app-types": "0.6.0"
+  },
   "deprecated": false,
   "description": "@firebase/database Types",
   "devDependencies": {
-    "typescript": "2.8.1"
+    "typescript": "3.8.3"
   },
   "files": [
     "index.d.ts"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "name": "@firebase/database-types",
-  "peerDependencies": {
-    "@firebase/app-types": "0.x"
-  },
   "repository": {
+    "directory": "packages/database-types",
     "type": "git",
-    "url": "https://github.com/firebase/firebase-js-sdk/tree/master/packages/database-types"
+    "url": "git+https://github.com/firebase/firebase-js-sdk.git"
   },
   "scripts": {
     "test": "tsc"
   },
-  "version": "0.3.2"
+  "version": "0.4.14"
 }
