@@ -34,46 +35,28 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "@firebase/database-types": "0.3.2",
-    "@firebase/logger": "0.1.1",
-    "@firebase/util": "0.2.2",
-    "faye-websocket": "0.11.1",
-    "tslib": "1.9.0"
+    "@firebase/auth-interop-types": "0.1.4",
+    "@firebase/component": "0.1.8",
+    "@firebase/database-types": "0.4.14",
+    "@firebase/logger": "0.2.0",
+    "@firebase/util": "0.2.43",
+    "faye-websocket": "0.11.3",
+    "tslib": "1.11.1"
   },
   "deprecated": false,
   "description": "This is the Firebase Realtime Database component of the Firebase JS SDK.",
   "devDependencies": {
-    "@types/chai": "4.1.2",
-    "@types/mocha": "5.0.0",
-    "@types/node": "9.6.4",
-    "@types/sinon": "4.3.1",
-    "chai": "4.1.2",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-firefox-launcher": "1.1.0",
-    "karma-mocha": "1.3.0",
-    "karma-sauce-launcher": "1.2.0",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "2.0.9",
-    "mocha": "5.0.5",
-    "npm-run-all": "4.1.2",
-    "nyc": "11.6.0",
-    "rollup": "0.57.1",
-    "rollup-plugin-commonjs": "9.1.0",
-    "rollup-plugin-node-resolve": "3.3.0",
-    "rollup-plugin-typescript2": "0.12.0",
-    "sinon": "4.5.0",
-    "source-map-loader": "0.2.3",
-    "ts-loader": "3.5.0",
-    "ts-node": "5.0.1",
-    "typescript": "2.8.1",
-    "webpack": "3.11.0"
+    "@firebase/app": "0.6.0",
+    "@firebase/app-types": "0.6.0",
+    "rollup": "2.0.6",
+    "rollup-plugin-typescript2": "0.26.0",
+    "typescript": "3.8.3"
   },
+  "esm2017": "dist/index.esm2017.js",
   "files": [
     "dist"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "main": "dist/index.node.cjs.js",
   "module": "dist/index.esm.js",
