@@ -1,28 +1,28 @@
 {
-  "_from": "@firebase/firestore-types@0.7.0",
-  "_id": "@firebase/firestore-types@0.7.0",
+  "_from": "@firebase/firestore-types@1.10.1",
+  "_id": "@firebase/firestore-types@1.10.1",
   "_inBundle": false,
-  "_integrity": "sha512-jyKRcKnSh3CSEPL4xGOZNoOXEiv7YmFK/JEcdd/4cAH17/Xo+Pk67gk1E648LRKh6QPghgNvzNTY5R10mKbQNw==",
+  "_integrity": "sha512-vyKdm+AYUFT8XeUX62IOqaqPFCs/mAMoSEsqIz9HnSVsqCw/IocNjtjSa+3M80kRw4V8fI7JI+Xz6Wg5VJXLqA==",
   "_location": "/@firebase/firestore-types",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@firebase/firestore-types@0.7.0",
+    "raw": "@firebase/firestore-types@1.10.1",
     "name": "@firebase/firestore-types",
     "escapedName": "@firebase%2ffirestore-types",
     "scope": "@firebase",
-    "rawSpec": "0.7.0",
+    "rawSpec": "1.10.1",
     "saveSpec": null,
-    "fetchSpec": "0.7.0"
+    "fetchSpec": "1.10.1"
   },
   "_requiredBy": [
     "/@firebase/firestore"
   ],
-  "_resolved": "https://registry.npmjs.org/@firebase/firestore-types/-/firestore-types-0.7.0.tgz",
-  "_shasum": "bded7892868cf6b189a121d0f0cec468f1609995",
-  "_spec": "@firebase/firestore-types@0.7.0",
-  "_where": "C:\\Users\\dirty\\workspaces\\chrismuzzin.info\\functions\\node_modules\\@firebase\\firestore",
+  "_resolved": "https://registry.npmjs.org/@firebase/firestore-types/-/firestore-types-1.10.1.tgz",
+  "_shasum": "bf018f9c495f470592de745389474dc1c2960d3f",
+  "_spec": "@firebase/firestore-types@1.10.1",
+  "_where": "C:\\Users\\dirty\\workspaces\\Website\\functions\\node_modules\\@firebase\\firestore",
   "author": {
     "name": "Firebase",
     "email": "firebase-support@google.com",
@@ -35,22 +35,24 @@
   "deprecated": false,
   "description": "@firebase/firestore Types",
   "devDependencies": {
-    "typescript": "2.8.1"
+    "typescript": "3.8.3"
   },
   "files": [
     "index.d.ts"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "name": "@firebase/firestore-types",
   "peerDependencies": {
     "@firebase/app-types": "0.x"
   },
   "repository": {
+    "directory": "packages/firestore-types",
     "type": "git",
-    "url": "https://github.com/firebase/firebase-js-sdk/tree/master/packages/firestore-types"
+    "url": "git+https://github.com/firebase/firebase-js-sdk.git"
   },
   "scripts": {
     "test": "tsc"
   },
-  "version": "0.7.0"
+  "version": "1.10.1"
 }
