@@ -41,56 +34,35 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "@firebase/firestore-types": "0.7.0",
-    "@firebase/logger": "0.1.1",
-    "@firebase/webchannel-wrapper": "0.2.11",
-    "grpc": "1.13.1",
-    "tslib": "1.9.0"
+    "@firebase/component": "0.1.8",
+    "@firebase/firestore-types": "1.10.1",
+    "@firebase/logger": "0.2.0",
+    "@firebase/util": "0.2.43",
+    "@firebase/webchannel-wrapper": "0.2.38",
+    "@grpc/proto-loader": "^0.5.0",
+    "grpc": "1.24.2",
+    "tslib": "1.11.1"
   },
   "deprecated": false,
   "description": "This is the [Cloud Firestore](https://firebase.google.com/docs/firestore/) component of the [Firebase JS SDK](https://www.npmjs.com/package/firebase).",
   "devDependencies": {
-    "@types/chai": "4.1.2",
-    "@types/long": "3.0.32",
-    "@types/mocha": "5.0.0",
-    "@types/sinon": "4.3.1",
-    "chai": "4.1.2",
-    "chai-as-promised": "7.1.1",
-    "indexeddbshim": "3.6.2",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-firefox-launcher": "1.1.0",
-    "karma-mocha": "1.3.0",
-    "karma-sauce-launcher": "1.2.0",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-spec-reporter": "0.0.32",
-    "karma-summary-reporter": "1.5.0",
-    "karma-webpack": "2.0.9",
-    "long": "3.2.0",
-    "mkdirp": "0.5.1",
-    "mocha": "5.0.5",
-    "npm-run-all": "4.1.2",
-    "nyc": "11.6.0",
-    "prettier": "1.12.0",
-    "rollup": "0.57.1",
-    "rollup-plugin-commonjs": "9.1.0",
-    "rollup-plugin-copy-assets": "1.0.0",
-    "rollup-plugin-node-resolve": "3.3.0",
-    "rollup-plugin-replace": "2.0.0",
-    "rollup-plugin-typescript2": "0.12.0",
-    "sinon": "4.5.0",
-    "source-map-loader": "0.2.3",
-    "ts-loader": "3.5.0",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.8.1",
-    "webpack": "3.11.0",
-    "yargs": "11.0.0"
+    "protobufjs": "6.8.9",
+    "rollup": "2.0.6",
+    "rollup-plugin-copy-assets": "1.1.0",
+    "rollup-plugin-json": "4.0.0",
+    "rollup-plugin-node-resolve": "5.2.0",
+    "rollup-plugin-replace": "2.2.0",
+    "rollup-plugin-sourcemaps": "0.5.0",
+    "rollup-plugin-terser": "5.3.0",
+    "rollup-plugin-typescript2": "0.26.0",
+    "ts-node": "8.6.2",
+    "typescript": "3.8.3"
   },
+  "esm2017": "dist/index.esm2017.js",
   "files": [
     "dist"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "main": "dist/index.node.cjs.js",
   "module": "dist/index.esm.js",
