@@ -1,28 +1,28 @@
 {
-  "_from": "@firebase/functions-types@0.2.0",
-  "_id": "@firebase/functions-types@0.2.0",
+  "_from": "@firebase/functions-types@0.3.16",
+  "_id": "@firebase/functions-types@0.3.16",
   "_inBundle": false,
-  "_integrity": "sha512-q1FB3YKEAnWd+FpIL5Xn0B1BXO2IowrAdrSViXkFxNZVpp9iCzQ8Ytcbr3V1xUr3dnmoW/V7zkZJZGuwBgiVhw==",
+  "_integrity": "sha512-kHhBvSYiY2prY4vNQCALYs1+OruTdylvGemHG6G6Bs/rj3qw7ui3WysBsDU/rInJitHIcsZ35qrtanoJeQUIXQ==",
   "_location": "/@firebase/functions-types",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@firebase/functions-types@0.2.0",
+    "raw": "@firebase/functions-types@0.3.16",
     "name": "@firebase/functions-types",
     "escapedName": "@firebase%2ffunctions-types",
     "scope": "@firebase",
-    "rawSpec": "0.2.0",
+    "rawSpec": "0.3.16",
     "saveSpec": null,
-    "fetchSpec": "0.2.0"
+    "fetchSpec": "0.3.16"
   },
   "_requiredBy": [
     "/@firebase/functions"
   ],
-  "_resolved": "https://registry.npmjs.org/@firebase/functions-types/-/functions-types-0.2.0.tgz",
-  "_shasum": "936d3f77957431a8ef1817975fee26464575802b",
-  "_spec": "@firebase/functions-types@0.2.0",
-  "_where": "C:\\Users\\dirty\\workspaces\\chrismuzzin.info\\functions\\node_modules\\@firebase\\functions",
+  "_resolved": "https://registry.npmjs.org/@firebase/functions-types/-/functions-types-0.3.16.tgz",
+  "_shasum": "be0362d7f61648fdf36a7d95de239eddee88f931",
+  "_spec": "@firebase/functions-types@0.3.16",
+  "_where": "C:\\Users\\dirty\\workspaces\\Website\\functions\\node_modules\\@firebase\\functions",
   "author": {
     "name": "Firebase",
     "email": "firebase-support@google.com",
@@ -35,19 +35,21 @@
   "deprecated": false,
   "description": "@firebase/functions Types",
   "devDependencies": {
-    "typescript": "2.8.1"
+    "typescript": "3.8.3"
   },
   "files": [
     "index.d.ts"
   ],
+  "homepage": "https://github.com/firebase/firebase-js-sdk#readme",
   "license": "Apache-2.0",
   "name": "@firebase/functions-types",
   "repository": {
+    "directory": "packages/functions-types",
     "type": "git",
-    "url": "https://github.com/firebase/firebase-js-sdk/tree/master/packages/functions-types"
+    "url": "git+https://github.com/firebase/firebase-js-sdk.git"
   },
   "scripts": {
     "test": "tsc"
   },
-  "version": "0.2.0"
+  "version": "0.3.16"
 }
