@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {FormGroup} from '@angular/forms';
+import { FormGroup } from '@angular/forms';
 import {FormValidatorService} from '../shared/services/form-validator.service';
 import {ToastrService} from 'ngx-toastr';
 import {AngularFireDatabase} from '@angular/fire/database';
